Move Wiremock-related helpers to their own dir
