+++ integration_tests/mockApis/tokenVerification.ts
@@ -1,5 +1,5 @@
 import { SuperAgentRequest } from 'superagent'
-import { stubFor } from './wiremock'
+import { stubFor } from '../../wiremock'
 
 export default {
   stubTokenVerificationPing: (status = 200): SuperAgentRequest =>
